LayoutTests/inspector/console/console-format-expected.txt

CONSOLE MESSAGE: line 12: Message format number %i, %d and %f
CONSOLE MESSAGE: line 13: Message %s for %s
CONSOLE MESSAGE: line 14: Object %o
CONSOLE MESSAGE: line 20: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 21: %o
CONSOLE MESSAGE: line 22: %O
CONSOLE MESSAGE: line 54: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 55: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 54: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 55: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 54: test
CONSOLE MESSAGE: line 55: test
CONSOLE MESSAGE: line 54: test named "test"
CONSOLE MESSAGE: line 55: test named "test"
CONSOLE MESSAGE: line 54: Error
CONSOLE MESSAGE: line 55: Error
CONSOLE MESSAGE: line 54: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 55: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 54: function () { return 1; }
CONSOLE MESSAGE: line 55: function () { return 1; }
CONSOLE MESSAGE: line 54: function () {
        return 2;
    }
CONSOLE MESSAGE: line 55: function () {
        return 2;
    }
CONSOLE MESSAGE: line 54: 0.12
CONSOLE MESSAGE: line 55: 0.12
CONSOLE MESSAGE: line 54: http://webkit.org/
CONSOLE MESSAGE: line 55: http://webkit.org/
CONSOLE MESSAGE: line 54: null
CONSOLE MESSAGE: line 55: 
CONSOLE MESSAGE: line 54: undefined
CONSOLE MESSAGE: line 55: 
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 55: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 55: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 55: [object Attr]
CONSOLE MESSAGE: line 54: [object Object]
CONSOLE MESSAGE: line 55: [object Object]
CONSOLE MESSAGE: line 54: NaN
CONSOLE MESSAGE: line 55: NaN
CONSOLE MESSAGE: line 54: Infinity
CONSOLE MESSAGE: line 55: Infinity
CONSOLE MESSAGE: line 54: -Infinity
CONSOLE MESSAGE: line 55: -Infinity
CONSOLE MESSAGE: line 54: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 55: test,test2,,,test4,,,,,
Tests that console logging dumps proper messages.

Message format number 1, 2 and 3.5 console-format.html:12
Message format for string console-format.html:13
Object Object console-format.html:14
Array[10] console-format.html:20
Array[10] console-format.html:21
Array[10] console-format.html:22
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i console-format.html:54
[/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\…?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i] console-format.html:55
globals[0]
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
/foo\\bar\sbaz/i console-format.html:54
[/foo\\bar\sbaz/i] console-format.html:55
globals[1]
/foo\\bar\sbaz/i
test console-format.html:54
["test"] console-format.html:55
globals[2]
"test"
test named "test" console-format.html:54
["test named "test""] console-format.html:55
globals[3]
"test named "test""
Error console-format.html:54
[Error] console-format.html:55
globals[4]
Error
<p id="p">Tests that console logging dumps proper messages.</p> console-format.html:54
[<p>] console-format.html:55
globals[5]
<p id="p">Tests that console logging dumps proper messages.</p>
function () { return 1; } console-format.html:54
[function () { return 1; }] console-format.html:55
globals[6]
function () { return 1; }
function () {
        return 2;
    } console-format.html:54
[function () {
        return 2;
    }] console-format.html:55
globals[7]
function () {
        return 2;
    }
0.12 console-format.html:54
[0.12] console-format.html:55
globals[8]
0.12
http://webkit.org/ console-format.html:54
["http://webkit.org/"] console-format.html:55
globals[9]
"http://webkit.org/"
null console-format.html:54
[null] console-format.html:55
globals[10]
null
undefined console-format.html:54
[undefined] console-format.html:55
globals[11]
undefined
attr="" console-format.html:54
[attr] console-format.html:55
globals[12]
attr=""
attr="value""value" console-format.html:54
[attr] console-format.html:55
globals[13]
attr="value""value"
id="x""x" console-format.html:54
[id] console-format.html:55
globals[14]
id="x""x"
Object console-format.html:54
[Object] console-format.html:55
globals[15]
Object
NaN console-format.html:54
[NaN] console-format.html:55
globals[16]
NaN
Infinity console-format.html:54
[Infinity] console-format.html:55
globals[17]
Infinity
-Infinity console-format.html:54
[-Infinity] console-format.html:55
globals[18]
-Infinity
["test", "test2", 4: "test4", foo: Object] console-format.html:54
[Array[10]] console-format.html:55
globals[19]
["test", "test2", undefined × 2, "test4", undefined × 5]
